Do example deployments to MLE and Sagemaker
In this commit we have added examples of the deployment process to
deploy the model to both AWS Sagemaker and GCP ML-Engine.  Deployment
packages are kept as generic as possible.  Data is also included in the
repository in this case since it is small and cleans up the notebook
code by removing embedded bash to curl the data files.